sep scalar and complex tests
