+++ test/08-ast-scalar-test.js
@@ -1,7 +1,5 @@
 let test = require('tape')
 let parse = require('../')
-let isVector = require('../src/ast/_is-vector')
-let isMap = require('../src/ast/_is-map')
 
 /**
  * ast nodes have the following required properties: type, line, column
@@ -33,6 +31,7 @@ let isMap = require('../src/ast/_is-map')
  *   - must contain at least one pragma
  *   - ex. { type: 'arcfile', values: [{ type: 'pragma', name: 'example', values: [] }] }
  * - pragma
+ *   - eats init newline
  *   - has name property
  *   - raw will contain full text including any inline comment
  *   - values can be empty and scalar types
@@ -42,18 +41,21 @@ let isMap = require('../src/ast/_is-map')
  *   - can only contain empty and scalar values
  *   - ex. { type: 'array', values: [{type: 'newline', value: '\n'}, {type:'number', value: 2}, {type:'string', value: 'hi'}]}
  * - vector
+ *   - eats initial newline
  *   - can only contain empty and scalar values
  *   - has name property
  *   - raw property will contain full text including comments
  *   - { type: 'vector', name: 'stuff', values: []}
  * - map
- *   - can only contain empty and key values
+ *   - eats initial newline
+ *   - can only contain empty, key and vector values
  *   - has name property
  *   - raw property will contain full text including comments
  *   - { type: 'map', name: 'cats', values: [{type: 'key', name: 'catID', values: []}]}
  * - key
+ *   - eat two spaces
  *   - can only belong to a map!
- *   - can only contain empty and scalar values
+ *   - can contain empty and scalar values
  *   - can also contain array, and vector values (this is the deepest nesting allowed by arcfiles)
  *   - if raw present it forces newline
  *   - has name property
@@ -188,164 +190,3 @@ one true 3 # comment2`
   console.dir(parsed, { depth: null })
   t.same(parsed, expected, 'successfully parsed array')
 })
-
-test('isVector', t => {
-  t.plan(1)
-  let mock = `
-@vector-test
-vec   # hi
-  one  # uh oh
-  2
-  # what about now
-  true
- some stuff here`
-  let tokens = parse.lexer(mock)
-  let start = tokens.slice(3, tokens.length) // start at 'vec'
-  t.ok(isVector(start), 'found a valid vector')
-})
-
-test('is not a vector', t => {
-  t.plan(1)
-  let mock = `
-@is-vec-test
-map
-  one two`
-  let tokens = parse.lexer(mock)
-  let start = tokens.slice(3, tokens.length) // start at 'map' token
-  t.ok(isVector(start) === false, 'did not find a valid vector')
-})
-
-test('ast vectors', t => {
-  t.plan(1)
-
-  let mock = `
-# comment1
-@pragma
-# comment2
-named # comment3
-  vector
-  of
-  # comment4
-  values
-this should be ignored`
-
-  let expected = {
-    type: 'arcfile',
-    values: [
-      { type: 'newline', value: '\n', line: 1, column: 1 },
-      { type: 'comment', value: '# comment1', line: 2, column: 1 },
-      { type: 'newline', value: '\n', line: 2, column: 11 },
-      {
-        type: 'pragma',
-        name: 'pragma',
-        raw: 'pragma',
-        line: 3,
-        column: 1,
-        values: [
-          { type: 'newline', value: '\n', line: 3, column: 8 },
-          { type: 'comment', value: '# comment2', line: 4, column: 1 },
-          { type: 'newline', value: '\n', line: 4, column: 11 },
-          {
-            type: 'vector',
-            name: 'named',
-            raw: 'named # comment3',
-            values: [
-              { type: 'space', value: ' ', line: 6, column: 1 },
-              { type: 'space', value: ' ', line: 6, column: 2 },
-              { type: 'string', value: 'vector', line: 6, column: 3 },
-              { type: 'newline', value: '\n', line: 6, column: 9 },
-              { type: 'space', value: ' ', line: 7, column: 1 },
-              { type: 'space', value: ' ', line: 7, column: 2 },
-              { type: 'string', value: 'of', line: 7, column: 3 },
-              { type: 'newline', value: '\n', line: 7, column: 5 },
-              { type: 'space', value: ' ', line: 8, column: 1 },
-              { type: 'space', value: ' ', line: 8, column: 2 },
-              {
-                type: 'comment',
-                value: '# comment4',
-                line: 8,
-                column: 3
-              },
-              { type: 'newline', value: '\n', line: 8, column: 13 },
-              { type: 'space', value: ' ', line: 9, column: 1 },
-              { type: 'space', value: ' ', line: 9, column: 2 },
-              { type: 'string', value: 'values', line: 9, column: 3 },
-              { type: 'newline', value: '\n', line: 9, column: 9 }
-            ]
-          },
-          {
-            type: 'array',
-            line: 10,
-            column: 1,
-            values: [
-              { type: 'string', value: 'this', line: 10, column: 1 },
-              { type: 'space', value: ' ', line: 10, column: 5 },
-              { type: 'string', value: 'should', line: 10, column: 6 },
-              { type: 'space', value: ' ', line: 10, column: 12 },
-              { type: 'string', value: 'be', line: 10, column: 13 },
-              { type: 'space', value: ' ', line: 10, column: 15 },
-              { type: 'string', value: 'ignored', line: 10, column: 16 }
-            ]
-          }
-        ]
-      }
-    ]
-  }
-  let parsed = parse.ast(parse.lexer(mock))
-  console.dir(parsed, { depth: null })
-  t.same(parsed, expected, 'successfully parsed vector')
-})
-
-test('isMap', t => {
-  t.plan(1)
-  let mock = `
-@hi
-map #cool cool
-# what about now!
-  one two # fun
-  three
-    four
-    five
-    six
-  # this is ok
-  seven
-    8`
-  let tokens = parse.lexer(mock)
-  t.ok(isMap(tokens.slice(3, tokens.length)))
-})
-
-test('map', t => {
-  t.plan(1)
-  let mock = `
-@map-test
-m1
-  one 1`
-  let tokens = parse.lexer(mock)
-  let ast = parse.ast(tokens)
-  t.ok(true)
-  console.dir(ast, { depth: null })
-})
-
-/*
-test.only('ast should parse maps', t => {
-  t.plan(1)
-
-  let mock = `
-# comment1
-@pragma
-# comment2
-mappy #comment3
-  k1 val
-  # comment4
-  one true #comment5
-  vec1 1 2 3
-  vec2 # comment6
-    one # comment7
-    # comment8
-    2
-    false`
-  let parsed = parse.ast(parse.lexer(mock))
-  console.dir(parsed, { depth: null })
-  t.ok(true)
-})
-*/
